Bug Fixes
When a user signs up, the suburb lat and long were initialized with a value of zero,this caused it not to call the google maps api (as it already thought it had all the data it needed). Added checks to see if either lat or long is null if it is, check the database to see if we have that suburb, if not get the details from Google Maps API.

When updating a customer or a service provider with an item in a container (i.e. a skill in their skill array) the container would be cleared. Changed it so the containers are copied over from there old data.