More ServiceRequest updates
Changed ServiceRequest endpoints so they no longer need to use IDs in the URL

Added new DTO for tradies accepting a ServiceRequest, only contains the ID of the ServiceRequest you wish to accept.

{
    "serviceRequestId":1
}

Moved the start/end date and time to when the customer first creates the ServiceRequest, it should also contain the ServiceRequest type

{
    "cost": 10000.00,
    "description":"Poo poo",
    "dateTimeRange": {
        "startDate": "2023-11-20",
        "startTime": "9:00am",
        "endDate":"2023-11-22",
        "endTime":"5:00pm",
        "serviceType": "POOL_CLEANING"

    }
}

Moved the logic for checking if the customer is a subscriber or a pay-on-demand to work out the price in the createServiceRequest method rather than the tradieAcceptsServiceRequest method.

Changed getServiceProviderDetailsFromServiceRequest and getCustomerDetailsFromServiceRequest methods to get user info from the JWT rather than specifying an ID in the URL